Semetar 1 Lesson 6/src/main/java/dev/josh/PasswordValidator.java
package dev.josh;

public class PasswordValidator {

    public boolean validate(String password) {
        if (password == null || password.length() < 5 || password.length() > 10) {
            return false;
        }

        boolean hasUppercase = false;
        boolean hasLowercase = false;
        boolean hasSpecialChar = false;
        boolean hasDigit = false;

        for (char c : password.toCharArray()) {
            if (Character.isUpperCase(c)) hasUppercase = true;
            else if (Character.isLowerCase(c)) hasLowercase = true;
            else if (Character.isDigit(c)) hasDigit = true;
            else if (!Character.isLetterOrDigit(c)) hasSpecialChar = true;
        }

        return hasUppercase && hasLowercase && hasSpecialChar && hasDigit;
    }
}